@@ -1,52 +1,48 @@
 {
-  "name": "bitcoin-tracker",
-  "version": "1.0.0",
-  "description": "Live crypto-currency price tracker",
-  "main": "index.js",
-  "scripts": {
-    "start": "webpack-dev-server",
-    "build": "webpack",
-    "test": "jest --setupTestFrameworkScriptFile --watchAll"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/chazmcgrill/bitcoin-tracker.git"
-  },
-  "keywords": [
-    "cryptocurrency",
-    "bitcoin",
-    "react"
-  ],
-  "author": "Charlie Taylor",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/chazmcgrill/bitcoin-tracker/issues"
-  },
-  "homepage": "https://github.com/chazmcgrill/bitcoin-tracker#readme",
+  "name": "coinage",
+  "version": "1.2.0",
+  "private": true,
   "dependencies": {
-    "react": "^16.4.0",
-    "react-dom": "^16.4.0"
+    "@types/jest": "24.0.23",
+    "@types/node": "12.12.7",
+    "@types/react-dom": "16.9.4",
+    "axios": "^0.19.0",
+    "node-sass": "^4.13.1",
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0",
+    "react-redux": "^7.1.3",
+    "react-scripts": "^3.4.1",
+    "redux": "^4.0.4",
+    "redux-thunk": "^2.3.0",
+    "typescript": "3.7.2"
+  },
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "echo \"Add tests please.\"",
+    "testwatch": "react-scripts test",
+    "eject": "react-scripts eject",
+    "preversion": "npm test",
+    "version": "git add .",
+    "postversion": "git push && git push --tags"
+  },
+  "eslintConfig": {
+    "extends": "react-app"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   },
   "devDependencies": {
-    "babel-core": "^6.26.3",
-    "babel-loader": "^7.1.4",
-    "babel-plugin-transform-object-rest-spread": "^6.26.0",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
-    "css-loader": "^0.28.11",
-    "enzyme": "^3.3.0",
-    "enzyme-adapter-react-16": "^1.1.1",
-    "extract-text-webpack-plugin": "^3.0.2",
-    "jest-cli": "^22.4.4",
-    "node-sass": "^4.9.0",
-    "sass-loader": "^6.0.7",
-    "style-loader": "^0.20.3",
-    "webpack": "^3.12.0",
-    "webpack-dev-server": "^2.11.2"
-  },
-  "jest": {
-    "setupFiles": [
-      "./src/setupTests.js"
-    ]
+    "@types/react": "^16.9.11",
+    "@types/react-redux": "^7.1.5"
   }
 }
