@@ -1,85 +1,99 @@
 import React, { Component } from 'react';
-import AddCoinListItem from './AddCoinListItem'
+import AddCoinListItem from './AddCoinListItem';
+import './ControlPanel.sass';
 
 class ControlPanel extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      selectCoins: null 
+    constructor(props) {
+        super(props);
+        this.state = {
+            selectCoins: null,
+        };
+        this.handleAddSubmit = this.handleAddSubmit.bind(this);
     }
-    this.handleAddSubmit = this.handleAddSubmit.bind(this);
-  }
 
-  componentWillReceiveProps(nextProps) {
-    const {selectCoins} = this.props;
-    if (nextProps.selectCoins !== selectCoins) {
-      this.setState({ selectCoins });
+    componentDidMount() {
+        const { selectCoins } = this.props;
+        this.setState({ selectCoins });
     }
-  }
-    
-  componentDidMount() {
-    const selectCoins = this.props.selectCoins;
-    this.setState({ selectCoins });
-  }
 
-  handleAddCoinClick(id) {
-    const selectCoins = this.state.selectCoins.map(c => (
-      c.id === id ? {...c, showing: !c.showing } : c
-    ));
-    this.setState({ selectCoins });
-  }
+    componentWillReceiveProps(nextProps) {
+        const { selectCoins } = this.props;
+        if (nextProps.selectCoins !== selectCoins) {
+            this.setState({ selectCoins });
+        }
+    }
 
-  handleAddSubmit() {
-    const selectCoins = this.state.selectCoins
-      .filter(c => c.showing)
-      .map(c => c.id);
-    this.props.handleAddCoins(selectCoins);
-  }
+    handleAddCoinClick(id) {
+        const { selectCoins } = this.state;
+        const newSelectCoins = selectCoins.map(c => (
+            c.id === id ? { ...c, showing: !c.showing } : c
+        ));
+        this.setState({ selectCoins: newSelectCoins });
+    }
 
-  render() {
-    const {addOpen} = this.props;
-    const form = addOpen ? (
-      <div>
-        <h4>choose coins to add...</h4>
-        <button
-          onClick={this.handleAddSubmit}
-          className="add-coins-submit"
-        >
-        Submit
-        </button>
-        <div className="add-coins"> 
-          {this.state.selectCoins.map(c => (
-            <AddCoinListItem 
-              handleAddCoinClick={() => this.handleAddCoinClick(c.id)} 
-              key={c.id} 
-              data={c} 
-            />
-          ))}
-        </div>
-        <button 
-          onClick={this.handleAddSubmit} 
-          className="add-coins-submit"
-        >
-        Submit
-        </button>
-      </div>
-    ) : null;
+    handleAddSubmit() {
+        const { selectCoins } = this.state;
+        const { handleAddCoins } = this.props;
+        const newSelectCoins = selectCoins
+            .filter(c => c.showing)
+            .map(c => c.id);
+        handleAddCoins(newSelectCoins);
+    }
 
-    return (
-      <div>
-        <div className="control-panel">    
-          <button 
-            className="add-button"
-            onClick={this.props.toggleAddOpen} >
-            Edit
-          </button>    
-          <button onClick={this.props.handleCurrency}>Currency</button>
-          <button onClick={this.props.handleRefresh}>Refresh</button>
-        </div>
-        {form}
-      </div>
-    )
-  }
+    render() {
+        const {
+            addOpen,
+            toggleAddOpen,
+            handleCurrency,
+            handleRefresh,
+        } = this.props;
+        const { selectCoins } = this.state;
+        const form = addOpen ? (
+            <div>
+                <h4>choose coins to add...</h4>
+                <button
+                    type="button"
+                    onClick={this.handleAddSubmit}
+                    className="add-coins-submit"
+                >
+                    Submit
+                </button>
+                <div className="add-coins">
+                    {selectCoins.map(c => (
+                        <AddCoinListItem
+                            handleAddCoinClick={() => this.handleAddCoinClick(c.id)}
+                            key={c.id}
+                            data={c}
+                        />
+                    ))}
+                </div>
+                <button
+                    type="button"
+                    onClick={this.handleAddSubmit}
+                    className="add-coins-submit"
+                >
+                    Submit
+                </button>
+            </div>
+        ) : null;
+
+        return (
+            <div>
+                <div className="control-panel">
+                    <button
+                        type="button"
+                        className="add-button"
+                        onClick={toggleAddOpen}
+                    >
+                        Edit
+                    </button>
+                    <button type="button" onClick={handleCurrency}>Currency</button>
+                    <button type="button" onClick={handleRefresh}>Refresh</button>
+                </div>
+                {form}
+            </div>
+        );
+    }
 }
 
 export default ControlPanel;
