@@ -1,101 +1,91 @@
 import React, { Component } from 'react';
+import { connect } from 'react-redux';
 import CoinList from './CoinList';
 import ControlPanel from './ControlPanel';
 import Footer from './Footer';
-import { getPrice } from '../getCoins';
-import { getData } from '../getCoinData';
-
-const FAVOURITES = [
-  'BTC', 'XRP', 'LTC', 'ETH', 'XMR', 
-  'ZEC', 'DSH', 'GNT', 'ADA', 'XVG',
-];
+import {
+    getCoinData,
+    getCoinPrice,
+    addCoins,
+    removeCoin,
+} from '../actions';
+import './BitcoinTracker.sass';
 
 class BitcoinTracker extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      coins: [
-        { id: 0, code: '...', name: "Loading", price: 0 },
-      ],
-      currDollar: true,
-      addOpen: false,
-      coinList: []
+    constructor(props) {
+        super(props);
+        this.state = {
+            currDollar: true,
+            addOpen: false,
+        };
     }
-    this.updateCoins = this.updateCoins.bind(this);
-    this.handleDelete = this.handleDelete.bind(this);
-    this.handleAddCoins = this.handleAddCoins.bind(this);
-  }
 
-  async updateCoins() {
-    const filtered = this.state.coinList.filter(a => a.showing);
-    const codes = filtered.map(c => c.code);
-    const prices = await getPrice(codes); 
-    const coins = filtered.map(c => {
-      return {...c, price: prices[c.code] };
-    });
-    this.setState({coins});
-  }
+    componentDidMount = async () => {
+        const { fetchCoinData } = this.props;
+        await fetchCoinData();
+        this.updateCoins();
+    }
 
-  async getCoinData() {
-    const coinData = await getData();
-    const coinList = Object.keys(coinData).map((coin, idx) => (
-      { 
-        id: idx, 
-        name: coinData[coin].CoinName,
-        imageURL: coinData[coin].ImageUrl,
-        code: coin,
-        showing: FAVOURITES.includes(coin)
-      }
-    ));
-    this.setState({ coinList });
-    this.updateCoins();
-  }
+    updateCoins = () => {
+        const { coinList, fetchCoinPrice } = this.props;
+        const codes = coinList.filter(c => c.showing).map(c => c.code);
+        fetchCoinPrice(codes);
+    }
 
-  handleAddCoins(ids) {
-    const coinList = this.state.coinList.map(c => (
-      ids.includes(c.id) ? { ...c, showing: true } : c
-    ));
-    this.setState({coinList}, () => this.updateCoins());
-  }
+    handleAddCoins = async (ids) => {
+        const { addCoinsHandler } = this.props;
+        await addCoinsHandler(ids);
+        this.updateCoins();
+    }
 
-  handleDelete(id) {
-    const coinList = this.state.coinList.map(c => (
-      c.id === id ? { ...c, showing: false } : c
-    ));
-    this.setState({ coinList }, () => this.updateCoins());
-  }
+    handleDelete = (id) => {
+        const { removeCoinHandler } = this.props;
+        removeCoinHandler(id);
+        this.updateCoins();
+    }
 
-  componentDidMount() {
-    this.getCoinData();
-  }
+    render() {
+        const { coinList } = this.props;
+        const { currDollar, addOpen } = this.state;
 
-  render() {
-    const coinList = this.state.coinList.filter(c => !c.showing);
-    const { coins, currDollar, addOpen } = this.state;
+        const addCoinList = coinList.filter(coin => !coin.showing);
+        const selectedCoins = coinList.filter(coin => coin.showing);
 
-    return (
-      <div>
-        <div className="container">
-          <h1>coinage</h1>
-          <CoinList 
-            coinData={coins}
-            currDollar={currDollar}
-            addOpen={addOpen}
-            handleDelete={this.handleDelete}
-          />
-          <ControlPanel 
-            selectCoins={coinList}
-            handleRefresh={this.updateCoins}
-            handleAddCoins={this.handleAddCoins}
-            addOpen={addOpen}
-            toggleAddOpen={() => this.setState({addOpen: !addOpen})}
-            handleCurrency={() => this.setState({currDollar: !currDollar})} 
-          />
-        </div>
-        <Footer />
-      </div>
-    )
-  }
+        return (
+            <div>
+                <div className="container">
+                    <h1>coinage</h1>
+                    <CoinList
+                        coinData={selectedCoins}
+                        currDollar={currDollar}
+                        addOpen={addOpen}
+                        handleDelete={this.handleDelete}
+                    />
+                    <ControlPanel
+                        selectCoins={addCoinList}
+                        handleRefresh={this.updateCoins}
+                        handleAddCoins={this.handleAddCoins}
+                        addOpen={addOpen}
+                        toggleAddOpen={() => this.setState({ addOpen: !addOpen })}
+                        handleCurrency={() => this.setState({ currDollar: !currDollar })}
+                    />
+                </div>
+                <Footer />
+            </div>
+        );
+    }
 }
 
-export default BitcoinTracker;
+const mapStateToProps = state => ({
+    errorMessage: state.coins.errorMessage,
+    coinList: state.coins.coins,
+});
+
+const mapDispatchToProps = dispatch => ({
+    fetchCoinData: () => dispatch(getCoinData()),
+    fetchCoinPrice: codes => dispatch(getCoinPrice(codes)),
+    addCoinsHandler: ids => dispatch(addCoins(ids)),
+    removeCoinHandler: id => dispatch(removeCoin(id)),
+});
+
+export default connect(mapStateToProps, mapDispatchToProps)(BitcoinTracker);
