@@ -1,7 +1,7 @@
 import requestData from './requestData';
 
 export function getData(codes) {
-  return requestData(codes)
-    .then(data => data)
-    .catch(error => error);
+    return requestData(codes)
+        .then(data => data)
+        .catch(error => error);
 }
