@@ -1,7 +1,7 @@
 import request from './request';
 
 export function getPrice(codes) {
-  return request(codes)
-    .then(data => data)
-    .catch(error => error);
+    return request(codes)
+        .then(data => data)
+        .catch(error => error);
 }
