@@ -1,8 +1,8 @@
-export default function requestData(codes) {
-  const url = 'https://min-api.cryptocompare.com/data/all/coinlist'
+export default function requestData() {
+    const url = 'https://min-api.cryptocompare.com/data/all/coinlist';
 
-  return fetch(url)
-    .then(response => response.json())
-    .then(data => data.Data)
-    .catch(error => error);
+    return fetch(url)
+        .then(response => response.json())
+        .then(data => data.Data)
+        .catch(error => error);
 }
